pyproject.toml
[project]
name = "packages"
version = "0.0.3"
description = ""
authors = [
    { name = "ljbcloud", email = "206458089+ljbcloud@users.noreply.github.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = '>=3.13,<4.0'
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "invoke (>=2.2.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
]

[build-system]
requires = ["poetry-core (>=2.0.0,<3.0.0)"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.ruff]
# enable pyflakes `E` and `F` codes by default
select = ["E", "F"]
ignore = ["E501", "E741"]
